+++ src/test/java/net/sf/beezle/sushi/io/BufferTest.java
@@ -77,7 +77,7 @@ public class BufferTest {
         Buffer buffer;
         InputStream src;
         ByteArrayOutputStream dest;
-        int length;
+        long length;
         byte[] found;
 
         buffer = new Buffer(1);
